+++ Libraries/ModelHelper.php
@@ -55,7 +55,7 @@ class ModelHelper {
 		$editFields = Field::getEditFields($config);
 
 		//if the model exists, load up the existing related items
-		if ($model->exists)
+		if ($model->exists && !$saving)
 		{
 			//make sure the relationships are loaded
 			foreach ($editFields['objectFields'] as $field => $info)
@@ -65,12 +65,16 @@ class ModelHelper {
 					//get all existing values for this relationship
 					if ($relatedItems = $model->{$field}()->get())
 					{
+						//the array that holds all the ids of the currently-related items
 						$relationsArray = array();
+
+						//the id-indexed array that holds all of the select option data for a relation.
+						//this holds the currently-related items and all of the available options
+						$autocompleteArray = array();
 
 						//iterate over the items
 						foreach ($relatedItems as $item)
 						{
-
 							//if this is a mutliple-value type (i.e. HasMany, HasManyAndBelongsTo), make sure this is an array
 							if ($info->multipleValues)
 							{
@@ -79,6 +83,12 @@ class ModelHelper {
 							else
 							{
 								$model->set_attribute($field, $item->{$item::$key});
+							}
+
+							//if this is an autocomplete field, we'll need to provide an array of arrays with 'id' and 'text' indexes
+							if ($info->autocomplete)
+							{
+								$autocompleteArray[$item->{$item::$key}] = array('id' => $item->{$item::$key}, 'text' => $item->{$info->nameField});
 							}
 						}
 
@@ -95,6 +105,12 @@ class ModelHelper {
 
 							//unset the relationships so we only get back what we need
 							$model->relationships = array();
+						}
+
+						//set the autocomplete array
+						if ($info->autocomplete)
+						{
+							$model->set_attribute($field.'_autocomplete', $autocompleteArray);
 						}
 					}
 					//if there are no values, then just set an empty array
@@ -359,7 +375,7 @@ class ModelHelper {
 		if ($selectedItems)
 		{
 			//if this isn't an array, set it up as one
-			$selectedItems = is_array($selectedItems) ? $selectedItems : array($selectedItems);
+			$selectedItems = is_array($selectedItems) ? $selectedItems : explode(',', $selectedItems);
 		}
 		else
 		{
@@ -401,13 +417,13 @@ class ModelHelper {
 			//set up the wheres
 			foreach ($info->searchFields as $search)
 			{
-				$query->or_where(DB::raw($search), 'LIKE', '%'.$term.'%');
+				$query->where(DB::raw($search), 'LIKE', '%'.$term.'%');
 			}
 
-			//include the currently-selected items if there are any
+			//exclude the currently-selected items if there are any
 			if (count($selectedItems))
 			{
-				$query->or_where_in($relatedModel->table().'.'.$relatedModel::$key, $selectedItems);
+				$query->where_not_in($relatedModel->table().'.'.$relatedModel::$key, $selectedItems);
 			}
 
 			//set up the limits
@@ -432,8 +448,8 @@ class ModelHelper {
 		return array_map(function($m) use ($info, $model)
 		{
 			return array(
-				$model::$key => $m->{$model::$key},
-				$info->nameField => $m->{$info->nameField},
+				'id' => $m->{$model::$key},
+				'text' => $m->{$info->nameField},
 			);
 		}, $eloquentResults);
 	}