+++ filters.php
@@ -15,7 +15,7 @@ Route::filter('add_assets', function()
 
 	$assets->add('jquery.ui.css', 'css/ui/jquery-ui-1.9.1.custom.min.css');
 	$assets->add('jquery.ui.timepicker.css', 'css/ui/jquery.ui.timepicker.css');
-	$assets->add('chosen_css', 'css/chosen.css');
+	$assets->add('select2_css', 'js/jquery/select2/select2.css');
 	$assets->add('jquery.lw-colorpicker', 'css/jquery.lw-colorpicker.css');
 	$assets->add('main_style', 'css/main.css');
 
@@ -27,9 +27,8 @@ Route::filter('add_assets', function()
 	//jquery core
 	$assets->add('jquery', 'js/jquery/jquery-1.8.2.min.js');
 
-	//jquery chosen and ajax chosen
-	$assets->add('jquery-chosen', 'js/jquery/jquery.chosen.min.js');
-	$assets->add('jquery-ajax-chosen', 'js/jquery/jquery.ajax-chosen.min.js');
+	//select2
+	$assets->add('jquery-select2', 'js/jquery/select2/select2.js');
 
 	//jquery ui
 	$assets->add('jquery.ui', 'js/jquery/jquery-ui-1.9.1.custom.min.js');