+++ changelog.md
@@ -6,8 +6,11 @@
 - You can now select nested belongs_to relationships in the columns array
 - Primary key fields are now hidden by default in the edit window unless you put them in your edit array
 - New languages: Spanish (es), Basque (eu), Dutch (nl), Polish (pl)
+- Moved from Chosen select boxes to Select2
+- Bugfix: Constraints on autocomplete fields now constrain the autocomplete search
 - Bugfix: Multiple constraints now work properly
 - Bugfix: Character limits on text fields no longer limit the string on every keystroke
+- Bugfix: Relationship options now sort by the name field
 
 ### 3.1.0
 - Localization support. Big thanks to [Andrew Dworn](https://github.com/andrewdworn) for all the work he put into this